BJS2-27340 balance_audit +

+++ src/main/resources/db/changelog/changeset/V005_BJS2-27340_balance_audit.sql
@@ -1,7 +1,7 @@
 CREATE TABLE balance_audit (
     id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY UNIQUE,
     account_id BIGINT NOT NULL,
-    balance_version INTEGER DEFAULT 0,
+    balance_version BIGINT DEFAULT 0,
     authorized_balance NUMERIC,
     actual_balance NUMERIC,
     operation_id BIGINT NOT NULL,
